+++ content/themes/dude/blocks/pricing-hero/src/index.js
@@ -1,25 +1,18 @@
 import { registerBlockType } from '@wordpress/blocks';
 import { useBlockProps, InnerBlocks } from '@wordpress/block-editor';
 import Edit from './edit';
+import deprecated from './deprecated';
 import './style.scss';
 
-registerBlockType( 'dude/pricing-hero', {
+registerBlockType('dude/pricing-hero', {
   edit: Edit,
-  save: ( { attributes } ) => {
-    const { paddingTopDesktop, paddingBottomDesktop, paddingTopMobile, paddingBottomMobile } = attributes;
-
-    const blockProps = useBlockProps.save( {
+  save: () => {
+    const blockProps = useBlockProps.save({
       className: 'block block-pricing-hero',
-      style: {
-        '--padding-top-desktop': `${ paddingTopDesktop }px`,
-        '--padding-bottom-desktop': `${ paddingBottomDesktop }px`,
-        '--padding-top-mobile': `${ paddingTopMobile }px`,
-        '--padding-bottom-mobile': `${ paddingBottomMobile }px`,
-      },
-    } );
+    });
 
     return (
-      <section { ...blockProps }>
+      <section {...blockProps}>
         <div className="container">
           <div className="content is-layout-grid">
             <InnerBlocks.Content />
@@ -28,4 +21,5 @@ registerBlockType( 'dude/pricing-hero', {
       </section>
     );
   },
-} );
+  deprecated,
+});